refactor: Modificando locais de arquivos e ajustes para nova estrutura
