--- media/style.css
@@ -1,120 +1,0 @@
-/* media/style.css */
-body {
-    font-family: sans-serif;
-    margin: 0;
-    padding: 10px;
-}
-
-.toolbar {
-    display: flex;
-    gap: 10px;
-    margin-bottom: 10px;
-}
-
-#tableContainer {
-    overflow: auto;
-}
-
-table {
-    border-collapse: collapse;
-    width: 100%;
-}
-
-th,
-td {
-    border: 1px solid #ccc;
-    padding: 8px;
-}
-
-button {
-    padding: 5px 10px;
-    background-color: #007BFF;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-/* Modal */
-.modal.hidden {
-    display: none;
-}
-
-.modal {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background: rgba(0, 0, 0, 0.5);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    z-index: 999;
-}
-
-.modal-content {
-    background: white;
-    padding: 30px;
-    width: 100%;
-    max-width: 400px;
-    border-radius: 10px;
-    position: relative;
-    box-shadow: 0px 5px 20px rgba(0, 0, 0, 0.2);
-}
-
-.modal-content h2 {
-    margin-top: 0;
-    font-size: 24px;
-    text-align: center;
-}
-
-.modal-content form label {
-    display: block;
-    margin-top: 15px;
-    font-weight: bold;
-}
-
-.modal-content form input {
-    width: 100%;
-    padding: 8px;
-    margin-top: 5px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-}
-
-.modal-content button {
-    margin-top: 20px;
-    width: 100%;
-    padding: 10px;
-    background: #4CAF50;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 16px;
-}
-
-.modal-content .btn-delete {
-    margin-top: 20px;
-    width: 100%;
-    padding: 10px;
-    background: #af4c4c;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 16px;
-}
-
-.modal-content button:hover {
-    background: #45a049;
-}
-
-.close-btn {
-    position: absolute;
-    top: 10px;
-    right: 15px;
-    font-size: 22px;
-    cursor: pointer;
-}
