+++ src/view/table/script.js
